refactor: clean up extra files and directories
