+++ SparkyFitnessMobile/src/services/CalorieWidgetBridge.ts
